remove avahi move python avahi client back into compute install numpy and urwid as apk packages to improve build time, update deps accordingly
