+++ .gitignore
@@ -10,6 +10,7 @@ __pycache__/
 .Python
 venv/
 env/
+.env/
 
 api/docs/build/
 develop-eggs/
