+++ Dockerfile
@@ -7,29 +7,24 @@ FROM resin/raspberrypi3-alpine:3.6
 # enable container init system.
 ENV INITSYSTEM on
 ENV RUNNING_ON_PI 1
-
+ENV DBUS_SYSTEM_BUS_ADDRESS=unix:path=/host/run/dbus/system_bus_socket
 ENV PYTHONPATH /data/packages/
 
 RUN apk add --update \
       dropbear \
-      avahi \
       gnupg \
-      gcc \
-      musl-dev \
-      make \
       nginx \
+      dbus \
       python3 \
-      python3-dev \
+      py3-urwid \
+      py3-numpy \
       && rm -rf /var/cache/apk/*
 
 COPY ./api /tmp/api
+COPY ./compute/alpine/avahi_tools /tmp/avahi_tools
 RUN pip3 install /tmp/api && \
-    rm -rf /tmp/api && \
-    apk del \
-      gcc \
-      musl-dev \
-      make \
-      python3-dev
+    pip3 install /tmp/avahi_tools && \
+    rm -rf /tmp/api
 
 RUN ln -s /usr/bin/python3 /usr/bin/python && \
     ln -s /usr/bin/pip3 /usr/bin/pip
@@ -43,17 +38,13 @@ COPY ./compute/alpine/services/updates /etc/init.d/
 COPY ./compute/alpine/services/networking /etc/init.d/
 COPY ./compute/alpine/services/dropbear /etc/init.d/
 
-COPY ./compute/alpine/services/avahi/api.service /etc/avahi/services/
-RUN rm /etc/avahi/services/ssh.service
-
 COPY ./compute/alpine/scripts/* /usr/local/bin/
 COPY ./compute/scripts/announce_mdns.py /usr/local/bin/
 
 RUN rc-update add api && \
     rc-update add nginx && \
     rc-update add updates && \
-    rc-update add dropbear && \
-    rc-update add avahi-daemon
+    rc-update add dropbear
 
 COPY ./compute/alpine/conf/rc.conf /etc/rc.conf
 COPY ./compute/alpine/conf/nginx.conf /etc/nginx/nginx.conf
